use old keep hashtags implementation :shrug:
