+++ src/features/keepTweetedHashtags.ts
@@ -1,12 +1,33 @@
 import {isHTMLElement} from '../helpers/domHelpers';
-import {onComposerDisabledStateChange, onComposerShown} from '../helpers/tweetdeckHelpers';
+import {onComposerShown} from '../helpers/tweetdeckHelpers';
+import {HandlerOf} from '../helpers/typeHelpers';
 import {makeBTDModule} from '../types/btdCommonTypes';
 
 export const keepTweetedHashtagsInComposer = makeBTDModule(({jq, settings}) => {
   if (!settings.saveTweetedHashtags) {
     return;
   }
-  let savedHashtags: string[] = [];
+  let hashtags: string[] = [];
+
+  // Save hashtags when typing.
+  document.body.addEventListener(
+    'keyup',
+    (ev) => {
+      if (
+        !isHTMLElement(ev.target) ||
+        !ev.target.matches('textarea.js-compose-text') ||
+        !(ev.target instanceof HTMLTextAreaElement)
+      ) {
+        return;
+      }
+
+      const tweetText = ev.target.value;
+      // @ts-expect-error
+      const tweetedHashtags = window.twttrTxt.extractHashtags(tweetText);
+      hashtags = tweetedHashtags;
+    },
+    true
+  );
 
   function pasteHashtags() {
     const tweetTextArea = document.querySelector<HTMLTextAreaElement>('textarea.js-compose-text');
@@ -15,51 +36,61 @@ export const keepTweetedHashtagsInComposer = makeBTDModule(({jq, settings}) => {
       return;
     }
 
-    if (savedHashtags.length === 0) {
+    if (hashtags.length === 0) {
       return;
     }
 
-    tweetTextArea.value = ` ${savedHashtags.map((t) => `#${t}`).join(' ')}`;
+    tweetTextArea.value = ` ${hashtags.map((t) => `#${t}`).join(' ')}`;
     tweetTextArea.selectionStart = 0;
     tweetTextArea.selectionEnd = 0;
     tweetTextArea.dispatchEvent(new Event('change'));
   }
 
-  jq(document).on('keypress', 'textarea.js-compose-text', (e) => {
-    if (!isHTMLElement(e.target)) {
+  // Re-instate hashtags when the composer is enabled again.
+  onComposerDisabledStateChange((isDisabled) => {
+    console.log({isDisabled});
+    if (isDisabled) {
       return;
     }
 
-    if (!(e.target instanceof HTMLTextAreaElement)) {
-      return;
-    }
-
-    const tweetText = e.target.value;
-    // @ts-expect-error
-    const tweetedHashtags = window.twttrTxt?.extractHashtags(tweetText) as string[] | undefined;
-
-    if (!tweetedHashtags) {
-      return;
-    }
-
-    savedHashtags = tweetedHashtags;
+    pasteHashtags();
   });
 
-  jq(document).one('dataColumnsLoaded', () => {
-    onComposerShown((isVisible) => {
-      if (!isVisible) {
-        return;
-      }
+  // Re-add hashtags when the composer is back.
+  onComposerShown((isVisible) => {
+    if (!isVisible) {
+      return;
+    }
 
-      pasteHashtags();
-    });
-
-    onComposerDisabledStateChange((isDisabled) => {
-      if (isDisabled) {
-        return;
-      }
-
-      pasteHashtags();
-    });
+    pasteHashtags();
   });
 });
+
+function onComposerDisabledStateChange(callback: HandlerOf<boolean>) {
+  const tweetComposerObserver = new MutationObserver(() => {
+    const tweetComposer = document.querySelector<HTMLTextAreaElement>(
+      '.drawer[data-drawer="compose"] textarea.js-compose-text'
+    );
+    callback(tweetComposer?.disabled ?? false);
+  });
+
+  onComposerShown((isVisible) => {
+    if (!isVisible) {
+      tweetComposerObserver.disconnect();
+      return;
+    }
+
+    const tweetComposer = document.querySelector(
+      '.drawer[data-drawer="compose"] textarea.js-compose-text'
+    );
+
+    if (!tweetComposer) {
+      return;
+    }
+
+    tweetComposerObserver.observe(tweetComposer, {
+      attributeFilter: ['disabled'],
+      attributes: true,
+    });
+  });
+}
